# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nutrients API",
        "description": "REST API for managing ingredients and their nutritional data.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/auth/validate-access-token": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Validate a bearer access token",
                "operationId": "cd2c295b52742448174746ee7ea4bbca",
                "responses": {
                    "200": {
                        "description": "Token is valid"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "503": {
                        "description": "Auth service unavailable"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout and invalidate tokens",
                "operationId": "f1edef7f79a0461455ab182363b1fd37",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "503": {
                        "description": "Auth service unavailable"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "frontend": {
                "type": "http",
                "description": "Bearer token obtained from POST /api/auth/login. Also requires X-Refresh-Token, X-Application-Name, and X-Client-Url headers on every request.",
                "scheme": "bearer"
            }
        }
    },
...
}